Move trait into behavior
This allows to keep API-specific logic completely out of the table model.
This way, API-specific code won’t even be loaded when browsing the website.
